Split WindowIdentifier logic and allow constructing wayland only types
Fixes #52
