+++ Cargo.toml
@@ -16,7 +16,8 @@ feature_gtk3 = ["gdk3x11", "gdk3wayland", "gtk3"]
 feature_gtk4 = ["gdk4x11", "gdk4wayland", "gtk4"]
 feature_pipewire = ["pw", "libc"]
 log = ["tracing"]
-raw_handle = ["raw-window-handle", "wayland-client", "wayland-protocols", "wayland-backend"]
+raw_handle = ["raw-window-handle", "wayland"]
+wayland = ["wayland-client", "wayland-protocols", "wayland-backend"]
 
 [dependencies]
 enumflags2 = "0.7"
